Moved membersPreview into its own file
