+++ packages/ui_v2/src/components/commonV2/Stepper/Preview.tsx
@@ -1,0 +1,71 @@
+import React, { FC } from "react";
+
+import { MembersForm } from "@dorgtech/daocreator-lib";
+
+import LineGraphic from "../LineGraphic";
+
+export const MembersPreview: FC<{
+  form: MembersForm;
+  tokenSymbol: string;
+}> = ({ form, tokenSymbol }) => {
+  const reputationConfig = {
+    showPercentage: false,
+    height: "0.5rem",
+    symbol: "REP",
+    dataKey: "reputation",
+    nameKey: "address"
+  };
+  const tokenConfig = {
+    showPercentage: false,
+    height: "0.5rem",
+    symbol: tokenSymbol,
+    dataKey: "tokens",
+    nameKey: "address"
+  };
+  let totalReputationAmount = 0;
+  let totalTokenAmount = 0;
+  form.toState().map(({ reputation, tokens }: any) => {
+    totalReputationAmount += reputation;
+    totalTokenAmount += tokens;
+    return null;
+  });
+  const numberOfMembers = form.$.length;
+  return (
+    <div style={styles.membersPreview}>
+      <p>
+        {numberOfMembers} Member{numberOfMembers > 1 && "s"}
+      </p>
+      <div style={{ width: "17.5em" }}>
+        <p>Reputation Distribution</p>
+        <LineGraphic
+          data={form.toState() as any} // Working with this type is weird
+          total={totalReputationAmount}
+          config={reputationConfig}
+          style={styles.lineGraphic}
+        />
+      </div>
+      {totalTokenAmount > 0 && (
+        <div style={{ paddingTop: "20px", width: "17.5em" }}>
+          <p>{tokenSymbol} Token Distribution</p>
+          <LineGraphic
+            data={form.toState() as any} // Working with this type is weird
+            total={totalTokenAmount}
+            config={tokenConfig}
+            style={styles.lineGraphic}
+          />
+        </div>
+      )}
+    </div>
+  );
+};
+
+const styles = {
+  lineGraphic: {
+    padding: "unset"
+  },
+
+  membersPreview: {
+    marginTop: 28,
+    paddingRight: "8rem"
+  }
+};
