+++ packages/ui_v2/src/components/commonV2/Stepper/index.tsx
@@ -9,30 +9,23 @@ import { MDBBtn, MDBRow, MDBCollapse, MDBIcon, MDBCol } from "mdbreact";
 
 import { UtilityButton } from "./UtilityButton";
 import { simpleOptionsSwitcher } from "../../utils";
-import LineGraphic from "../LineGraphic";
+import { MembersPreview } from "./Preview";
 
 interface Props {
   index: number;
   form: DAOForm | DAOConfigForm | MembersForm | SchemesForm;
   Component: FC;
   title: string;
-  callbacks: any | undefined;
+  callbacks: any;
   step: number;
   launching: boolean;
-  daoName?: string;
 }
 
-interface ModalButtonProps {
+const ModalButton: FC<{
   step: number;
   index: number;
   setModal: any;
-}
-
-const ModalButton: FC<ModalButtonProps> = ({
-  step,
-  index,
-  setModal
-}: ModalButtonProps) => {
+}> = ({ step, index, setModal }) => {
   if (step === 1 && index === 1) {
     return <UtilityButton title={"Advanced"} openModal={setModal} />;
   } else if (step === 2 && index === 2) {
@@ -41,9 +34,7 @@ const ModalButton: FC<ModalButtonProps> = ({
   return <></>;
 };
 
-// const simpleConfigText: FC<SimpleConfigText> = ({ form }) => {
 const SimpleConfigTextform: FC<{ form: SchemesForm }> = ({ form }) => {
-  console.log(form);
   const simpleOptions = simpleOptionsSwitcher(form, true);
   const noDuplicateSimpleOptions = simpleOptions.slice(
     0,
@@ -68,61 +59,6 @@ const SimpleConfigTextform: FC<{ form: SchemesForm }> = ({ form }) => {
             </p>
           </div>
         )
-      )}
-    </div>
-  );
-};
-
-const MembersPreview: FC<{
-  form: MembersForm;
-  tokenSymbol: string;
-}> = ({ form, tokenSymbol }) => {
-  const reputationConfig = {
-    showPercentage: false,
-    height: "0.5rem",
-    symbol: "REP",
-    dataKey: "reputation",
-    nameKey: "address"
-  };
-  const tokenConfig = {
-    showPercentage: false,
-    height: "0.5rem",
-    symbol: tokenSymbol,
-    dataKey: "tokens",
-    nameKey: "address"
-  };
-  let totalReputationAmount = 0;
-  let totalTokenAmount = 0;
-  form.toState().map(({ reputation, tokens }: any) => {
-    totalReputationAmount += reputation;
-    totalTokenAmount += tokens;
-    return null;
-  });
-  const numberOfMembers = form.$.length;
-  return (
-    <div style={styles.membersPreview}>
-      <p>
-        {numberOfMembers} Member{numberOfMembers > 1 && "s"}
-      </p>
-      <div style={{ width: "17.5em" }}>
-        <p>Reputation Distribution</p>
-        <LineGraphic
-          data={form.toState() as any} // Working with this type is weird
-          total={totalReputationAmount}
-          config={reputationConfig}
-          style={styles.lineGraphic}
-        />
-      </div>
-      {totalTokenAmount > 0 && (
-        <div style={{ paddingTop: "20px", width: "17.5em" }}>
-          <p>{tokenSymbol} Token Distribution</p>
-          <LineGraphic
-            data={form.toState() as any} // Working with this type is weird
-            total={totalTokenAmount}
-            config={tokenConfig}
-            style={styles.lineGraphic}
-          />
-        </div>
       )}
     </div>
   );
@@ -322,12 +258,5 @@ const styles = {
   },
   schemePreview: {
     marginTop: 28
-  },
-  membersPreview: {
-    marginTop: 28,
-    paddingRight: "8rem"
-  },
-  lineGraphic: {
-    padding: "unset"
   }
 };
